pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coverup-pdf"
version = "0.4.0"
description = "A tool for redacting PDF files and images"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "Björn Seipel", email = "support@digidigital.de"}
]
maintainers = [
    {name = "Björn Seipel", email = "support@digidigital.de"}
]
keywords = ["pdf", "redaction", "privacy", "image", "editor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Editors",
    "Topic :: Office/Business",
    "Topic :: Security",
]
requires-python = ">=3.9"
dependencies = [
    "FreeSimpleGUI>=5.0.0",
    "pypdfium2>=4.0.0",
    "Pillow>=10.0.0",
    "fpdf2>=2.7.0",
    "appdirs>=1.4.4",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=5.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://coverup.digidigital.de"
Repository = "https://github.com/digidigital/CoverUP"
Issues = "https://github.com/digidigital/CoverUP/issues"

[project.scripts]
coverup = "coverup.main:main"

[project.gui-scripts]
coverup-gui = "coverup.main:main"


[project.entry-points.pyinstaller40]
hook-dirs = "coverup.__pyinstaller:get_hook_dirs"


[tool.setuptools]
packages = ["coverup", "coverup.pyinstaller_hooks"]

[tool.setuptools.package-data]
coverup = ["fonts/*.ttf", "fonts/*.otf"]
"coverup.pyinstaller_hooks" = ["*.py"]
